chore(project): rename file to make api docs read better
