+++ lib/routes/index.js
@@ -32,7 +32,7 @@ module.exports = function (
     push
   )
   const devicesImpl = require('../devices')(log, db, push)
-  const devicesSessions = require('./devices-sessions')(log, db, config, customs, push, devicesImpl)
+  const devicesSessions = require('./devices-and-sessions')(log, db, config, customs, push, devicesImpl)
   const emails = require('./emails')(log, db, mailer, config, customs, push)
   const password = require('./password')(
     log,
